Keep track of previous commands
Previously, the simulation only updated the joint states for those joints that received a command in the new message.
This caused a problem as the PTU was streaming continuously but the driving joints were only receving sporadically messages.
Now the joint commands are saved into a dict and overwriting old commands, but all the joint commands are processed so all joints that receive a message will be simulated properly.